ai-service/requirements.txt
Flask
Pillow
numpy
tensorflow
opencv-python-headless
